requirements.txt
os
urllib.parse
requests
json
datetime
tqdm
pydrive.auth
pydrive.drive